requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.24.0
pydantic==2.7.0
pydantic-settings==2.5.2
pyjwt==2.8.0
cryptography==41.0.7
httpx==0.25.2
asyncpg==0.29.0
pgvector==0.2.4
numpy==1.26.2
python-multipart==0.0.6
python-dotenv==1.0.0
redis==5.0.1
anthropic==0.8.0
openai==1.6.0